+++ src/Syntax/Resolve/Scope.hs
@@ -44,7 +44,7 @@ builtinScope = Scope
                                   , "<", ">", ">=", "<=", "==", "<>"
                                   , "||", "&&"
                                   , "@@", "lazy", "force" ]
-               , tyScope  =  build [ "int", "string", "bool", "unit", "float", "lazy" ]
+               , tyScope    = Map.insert (Name "lazy") (SVar (TgName "lazy" (-34))) $ build [ "int", "string", "bool", "unit", "float" ]
                , tyvarScope = mempty
                , modStack = []
                }
