+++ tests/gadt/gadt01.out
@@ -1,5 +1,5 @@
 gadt01.ml[3:16 ..3:23]: error
-  Could not match type type with 'e -> type
+  Expected a type, but this annotation is of kind 'e -> type
   Have you applied a type constructor to the wrong number of arguments?
 
   Arising in the declaration
