+++ tests/gadt/gadt07-escape-fail.out
@@ -1,7 +1,7 @@
 gadt07-escape-fail.ml[10:3 ..10:17]: error
   Could not match rigid type variable a with the type int
   • Note: the variable a was rigidified because it is an existential,
-  bound by the type of MkHid, (exp 'a * exp 'a) -> hidden
+          bound by the type of MkHid, exp 'a * exp 'a -> hidden
 
 
   Arising in the expression
