+++ tests/gadt/gadt09-pass.out
@@ -5,7 +5,7 @@ Int : {'a : type}. ('a ~ int) ⊃ repr 'a
 Some : {'a : type}. 'a -> option 'a
 None : {'a : type}. option 'a
 <$> : {'a : type}. {'a : type}. ('a -> 'a) -> option 'a -> option 'a
-<*> : {'a : type}. {'a : type}. option ('a -> 'a) -> option 'a -> option 'a
+<*> : {'a : type}. {'hg : type}. option ('hg -> 'a) -> option 'hg -> option 'a
 cast : {'a : type}. {'b : type}. repr 'a -> repr 'b -> 'a -> option 'b
 main : unit
 repr : type -> type
