+++ tests/gadt/pass_vector.out
@@ -3,7 +3,7 @@ S : natural -> natural
 Nil : {'n : natural}. {'a : type}. ('n ~ Z) ⊃ vect 'n 'a
 Cons : {'n : natural}. {'a : type}. {'k : natural}. ('n ~ S 'k) ⊃ ('a * vect 'k 'a) -> vect 'n 'a
 head : {'ez : type}. {'k : natural}. vect (S 'k) 'ez -> 'ez
-tail : {'a : type}. {'k : natural}. vect (S 'k) 'a -> vect 'k 'a
+tail : {'k : natural}. {'ec : type}. vect (S 'k) 'ec -> vect 'k 'ec
 main : int
 natural : type
 vect : natural -> type -> type
