+++ tests/rankn/impredicative01-fail.out
@@ -1,5 +1,5 @@
 impredicative01-fail.ml[6:38 ..6:52]: error
-  Could not match type {'a : type}. 'a -> 'a with 'ae -> 'ae
+  Could not match type forall 'a. 'a -> 'a with 'af -> 'af
   Have you applied a function to the wrong number of arguments?
 
   Arising in the expression
