+++ tests/rankn/tuple02-fail.out
@@ -1,6 +1,7 @@
 tuple02-fail.ml[1:41 ..1:56]: error
   Could not match rigid type variable a with the type int
-  • Note: the variable a was rigidified because of a subsumption constraint relating int -> unit with {'a : type}. 'a -> 'a
+  • Note: the variable a was rigidified because of a requirement that int -> unit
+          be as polymorphic as forall 'a. 'a -> 'a
 
 
   Arising in the expression
