+++ tests/types/fail_skolem_escape.out
@@ -1,7 +1,7 @@
 fail_skolem_escape.ml[2:19 ..2:38]: error
-  Rigid type variable a has escaped its scope of {'a : type}. 'a -> 'o
-  • Note: the variable a was rigidified because of a type ascription
-          against the type {'a : type}. 'a -> 'o,
+  Rigid type variable a has escaped its scope of forall 'a. 'a -> 'o
+  • Note: the variable a was rigidified because of an ascription against the type
+          forall 'a. 'a -> 'o,
 
   • Note: in type `'a`
   • Note: in the inferred type for escape_fail
