+++ tests/types/fail_skolem_pure.out
@@ -1,7 +1,7 @@
 fail_skolem_pure.ml[2:32 ..2:32]: error
   Could not match rigid type variable a with the type 'f 'a
-  • Note: the variable a was rigidified because of a type ascription
-  against the type {'a : type}. 'a -> 'f 'a
+  • Note: the variable a was rigidified because of an ascription against the type
+          forall 'a. 'a -> 'f 'a
 
 
   Arising in the expression
