+++ tests/visinst/visible01-fail.out
@@ -1,5 +1,5 @@
 visible01-fail.ml[4:23 ..4:23]: error
-  Could not match type unit with int
+  Could not match type int with unit
 
   Arising in the expression
   │
