+++ tests/visinst/visible05-pass.out
@@ -6,7 +6,7 @@ SZ : {'n : nat}. ('n ~ Z) ⊃ natural 'n
 SS : {'n : nat}. {'k : nat}. ('n ~ S 'k) ⊃ natural 'k -> natural 'n
 Flip : {'f : 'fv -> 'fu -> type}. {'a : 'fu}. {'b : 'fv}. 'f 'b 'a -> flip 'f 'a 'b
 getFlip : {'f : 'hq -> 'hu -> type}. {'a : 'hu}. {'b : 'hq}. flip 'f 'a 'b -> 'f 'b 'a
-@ : {'ic : type}. {'id : type}. {'iu : type}. ('iu -> 'id) -> ('ic -> 'iu) -> 'ic -> 'id
+@ : {'ic : type}. {'id : type}. {'iq : type}. ('iq -> 'id) -> ('ic -> 'iq) -> 'ic -> 'id
 ind : {'n : nat}. ('prop : nat -> type) -> 'prop Z -> ({'n : nat}. 'prop 'n -> 'prop (S 'n)) -> natural 'n -> 'prop 'n
 Nil : {'n : nat}. {'a : type}. ('n ~ Z) ⊃ vect 'n 'a
 Cons : {'n : nat}. {'a : type}. {'k : nat}. ('n ~ S 'k) ⊃ ('a * vect 'k 'a) -> vect 'n 'a
